Add files via upload
Fixed standardization of features and dropped features that may be 'unusable' for training. Increased complexity of NN model and lowered learning rate to help with consistency and reduce fluctuations in training. Explored dropping less correlated variables with the conclusion that it did not significantly affect model performance at best, and decreased performance slightly on average.